+++ src/ts/components/subtitleoverlay.ts
@@ -12,37 +12,43 @@ export class SubtitleOverlay extends Container<ContainerConfig> {
 
   private static readonly CLASS_CONTROLBAR_VISIBLE = 'controlbar-visible';
 
-  /**
-   * Inner label that renders the subtitle text
-   */
-  private subtitleLabel: Label<LabelConfig>;
-
   constructor(config: ContainerConfig = {}) {
     super(config);
 
-    this.subtitleLabel = new Label<LabelConfig>({ cssClass: 'ui-subtitle-label' });
-
     this.config = this.mergeConfig(config, {
       cssClass: 'ui-subtitle-overlay',
-      components: [this.subtitleLabel]
     }, this.config);
   }
 
   configure(player: bitmovin.player.Player, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
+    let subtitleManager = new ActiveSubtitleManager();
+
     player.addEventHandler(player.EVENT.ON_CUE_ENTER, (event: SubtitleCueEvent) => {
-      this.subtitleLabel.setText(event.text);
+      let labelToAdd = subtitleManager.cueEnter(event);
+
+      this.addComponent(labelToAdd);
+      this.updateComponents();
+
       this.show();
     });
     player.addEventHandler(player.EVENT.ON_CUE_EXIT, (event: SubtitleCueEvent) => {
-      this.hide();
-      this.subtitleLabel.setText('');
+      let labelToRemove = subtitleManager.cueExit(event);
+
+      this.removeComponent(labelToRemove);
+      this.updateComponents();
+
+      if (!subtitleManager.hasCues) {
+        this.hide();
+      }
     });
 
     let subtitleClearHandler = () => {
       this.hide();
-      this.subtitleLabel.setText('');
+      subtitleManager.clear();
+      this.removeComponents();
+      this.updateComponents();
     };
 
     player.addEventHandler(player.EVENT.ON_AUDIO_CHANGED, subtitleClearHandler);
@@ -64,5 +70,100 @@ export class SubtitleOverlay extends Container<ContainerConfig> {
 
     // Init
     subtitleClearHandler();
+  }
+}
+
+interface ActiveSubtitleCue {
+  event: SubtitleCueEvent;
+  label: SubtitleLabel;
+}
+
+interface ActiveSubtitleCueMap {
+  [id: string]: ActiveSubtitleCue;
+}
+
+class SubtitleLabel extends Label<LabelConfig> {
+
+  constructor(config: LabelConfig = {}) {
+    super(config);
+
+    this.config = this.mergeConfig(config, {
+      cssClass: 'ui-subtitle-label'
+    }, this.config);
+  }
+}
+
+class ActiveSubtitleManager {
+
+  private activeSubtitleCueMap: ActiveSubtitleCueMap;
+
+  constructor() {
+    this.activeSubtitleCueMap = {};
+  }
+
+  /**
+   * Calculates a unique ID for a subtitle cue, which is needed to associate an ON_CUE_ENTER with its ON_CUE_EXIT
+   * event so we can remove the correct subtitle in ON_CUE_EXIT when multiple subtitles are active at the same time.
+   * The start time plus the text should make a unique identifier, and in the only case where a collision
+   * can happen, two similar texts will displayed at a similar time so it does not matter which one we delete.
+   * The start time should always be known, because it is required to schedule the ON_CUE_ENTER event. The end time
+   * must not necessarily be known and therefore cannot be used for the ID.
+   * @param event
+   * @return {string}
+   */
+  private static calculateId(event: SubtitleCueEvent): string {
+    return event.start + event.text;
+  }
+
+  /**
+   * Adds a subtitle cue to the manager and returns the label that should be added to the subtitle overlay.
+   * @param event
+   * @return {SubtitleLabel}
+   */
+  cueEnter(event: SubtitleCueEvent): SubtitleLabel {
+    let id = ActiveSubtitleManager.calculateId(event);
+
+    let label = new SubtitleLabel({
+      text: event.text
+    });
+
+    this.activeSubtitleCueMap[id] = { event: event, label: label};
+
+    return label;
+  }
+
+  /**
+   * Removes the subtitle cue from the manager and returns the label that should be removed from the subtitle overlay.
+   * @param event
+   * @return {SubtitleLabel}
+   */
+  cueExit(event: SubtitleCueEvent): SubtitleLabel {
+    let id = ActiveSubtitleManager.calculateId(event);
+    let activeSubtitleCue = this.activeSubtitleCueMap[id];
+    delete this.activeSubtitleCueMap[id];
+    return activeSubtitleCue.label;
+  }
+
+  /**
+   * Returns the number of active subtitle cues.
+   * @return {number}
+   */
+  get cueCount(): number {
+    return Object.keys(this.activeSubtitleCueMap).length;
+  }
+
+  /**
+   * Returns true if there are active subtitle cues, else false.
+   * @return {boolean}
+   */
+  get hasCues(): boolean {
+    return this.cueCount > 0;
+  }
+
+  /**
+   * Removes all subtitle cues from the manager.
+   */
+  clear(): void {
+    this.activeSubtitleCueMap = {};
   }
 }