+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Add `Container#removeComponents()` to remove all child components of a container
+- Display multiple subtitle cues in parallel in `SubtitleOverlay`
 
 ## [2.1.1]